safari-600.7-branch/LayoutTests/fast/dom/shadow/offset-parent-does-not-leak-ua-shadow-expected.txt

offsetParent should not leak nodes in user agent Shadow DOM.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS child.offsetParent is shadow.positionedElement
PASS child.offsetParent is container
PASS child.offsetParent is null
PASS child.offsetParent is container
PASS successfullyParsed is true

TEST COMPLETE
